Add /support and /report API endpoints
