+++ internal/handlers.go
@@ -36,7 +36,6 @@ import (
 	"github.com/prologic/twtxt"
 	"github.com/prologic/twtxt/internal/session"
 	"github.com/prologic/twtxt/types"
-	"github.com/steambap/captcha"
 )
 
 const (
@@ -2446,71 +2445,6 @@ func (s *Server) SyndicationHandler() httprouter.Handle {
 	}
 }
 
-// SupportHandler ...
-func (s *Server) SupportHandler() httprouter.Handle {
-	return func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
-		ctx := NewContext(s.config, s.db, r)
-
-		if r.Method == "GET" {
-			ctx.Title = "Contact support"
-			s.render("support", w, ctx)
-			return
-		}
-
-		name := strings.TrimSpace(r.FormValue("name"))
-		email := strings.TrimSpace(r.FormValue("email"))
-		subject := strings.TrimSpace(r.FormValue("subject"))
-		message := strings.TrimSpace(r.FormValue("message"))
-
-		captchaInput := strings.TrimSpace(r.FormValue("captchaInput"))
-
-		// Get session
-		sess := r.Context().Value(session.SessionKey)
-		if sess == nil {
-			log.Warn("no session found")
-			ctx.Error = true
-			ctx.Message = fmt.Sprintf("no session found, do you have cookies disabled?")
-			s.render("error", w, ctx)
-			return
-		}
-
-		// Get captcha text from session
-		captchaText, isCaptchaTextAvailable := sess.(*session.Session).Get("captchaText")
-		if isCaptchaTextAvailable == false {
-			log.Warn("no captcha provided")
-			ctx.Error = true
-			ctx.Message = "no captcha text found"
-			s.render("error", w, ctx)
-			return
-		}
-
-		if captchaInput != captchaText {
-			log.Warn("incorrect captcha")
-			ctx.Error = true
-			ctx.Message = "Unable to match captcha text. Please try again."
-			s.render("error", w, ctx)
-			return
-		}
-
-		if err := SendSupportRequestEmail(s.config, name, email, subject, message); err != nil {
-			log.WithError(err).Errorf("unable to send support email for %s", email)
-			ctx.Error = true
-			ctx.Message = "Error sending support message! Please try again."
-			s.render("error", w, ctx)
-			return
-		}
-
-		log.Infof("support message email sent for %s", email)
-
-		ctx.Error = false
-		ctx.Message = fmt.Sprintf(
-			"Thank you for your message! Pod operator %s will get back to you soon!",
-			s.config.AdminName,
-		)
-		s.render("error", w, ctx)
-	}
-}
-
 // PodAvatarHandler ...
 func (s *Server) PodAvatarHandler() httprouter.Handle {
 	return func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
@@ -2633,34 +2567,6 @@ func (s *Server) TransferFeedHandler() httprouter.Handle {
 			ctx.Error = false
 			ctx.Message = "Feed ownership changed successfully."
 			s.render("error", w, ctx)
-		}
-	}
-}
-
-// CaptchaHandler ...
-func (s *Server) CaptchaHandler() httprouter.Handle {
-	return func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
-		img, err := captcha.NewMathExpr(150, 50)
-		if err != nil {
-			log.WithError(err).Errorf("unable to get generate captcha image")
-			http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-			return
-		}
-
-		// Save captcha text in session
-		sess := r.Context().Value(session.SessionKey)
-		if sess == nil {
-			log.Warn("no session found")
-			http.Error(w, "Unauthorized", http.StatusUnauthorized)
-			return
-		}
-		sess.(*session.Session).Set("captchaText", img.Text)
-
-		w.Header().Set("Content-Type", "image/png")
-		if err := img.WriteImage(w); err != nil {
-			log.WithError(err).Errorf("error sending captcha image repsonse")
-			http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-			return
 		}
 	}
 }
